ストアファイルを分割
